chore: split Commands into smaller components
