+++ app/common/renderer/components/SessionInspector/CommandsTab/CommandResult/CommandResult.module.css
@@ -1,0 +1,22 @@
+.commandResultModal :global(.ant-modal-footer) {
+  text-align: unset;
+}
+
+.commandResultModalOkButtonCol {
+  display: flex;
+  justify-content: flex-end;
+}
+
+.commandResultRawPreBlock {
+  margin: 0;
+}
+
+.commandResultTable :global(.ant-table-tbody > tr > td) {
+  vertical-align: top;
+}
+
+.commandResultTableCell {
+  margin: 0;
+  white-space: pre-wrap;
+  cursor: pointer;
+}
